+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # v0.1
 
-```react-gallery-carousel``` [![NPM](https://img.shields.io/npm/v/react-gallery-carousel.svg)](https://www.npmjs.com/package/react-gallery-carousel)
+```react-gallery-carousel```
 
-## v0.1.4 (2021-05-02)
+[![NPM](https://img.shields.io/npm/v/react-gallery-carousel.svg)](https://www.npmjs.com/package/react-gallery-carousel)
+
+## v0.1.5 (2021-05-02)
 
 ### Bug Fixes
 
@@ -14,9 +16,11 @@
 
 ### Enhancements
 
+- Allowed the size of the thumbnails to be customized.
+  ([Issue #11](https://github.com/yifaneye/react-gallery-carousel/issues/11))
+
 - Added handlers and listeners to the carousel.
   ([Issue #12](https://github.com/yifaneye/react-gallery-carousel/issues/12))
-
 
 ## v0.1.4 (2021-04-18)